C++_Python/center/requirements.txt
# Center Server Requirements

# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
paho-mqtt==1.6.1
websockets==12.0
pydantic==2.5.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
# or use SQLite (built-in)

# Development
python-multipart==0.0.6
aiofiles==23.2.1
python-dotenv==1.0.0
